+++ Plugins/Multiplayer/Source/NexusMultiplayer/Public/NTextRenderComponent.h
@@ -10,7 +10,7 @@
 
 
 
-UCLASS()
+UCLASS(meta = (BlueprintSpawnableComponent))
 class NEXUSMULTIPLAYER_API UNTextRenderComponent final : public UTextRenderComponent
 {
 	GENERATED_BODY()
@@ -24,20 +24,29 @@ public:
 	virtual void GetLifetimeReplicatedProps(TArray<FLifetimeProperty>& OutLifetimeProps) const override;
 	virtual void BeginPlay() override;
 	
-	UPROPERTY(Replicated, EditAnywhere, BlueprintReadWrite, Category = "Replicated", ReplicatedUsing=OnRep_TextValue)
+	UPROPERTY(Replicated, ReplicatedUsing=OnRep_TextValue)
 	FString CachedValue;
 
 	UFUNCTION()
 	void OnRep_TextValue();
 	
-	UFUNCTION(Server, Reliable, BlueprintCallable, Category = "NEXUS")
-	void SetFromName(const FName& NewValue);
+	UFUNCTION(Server, Reliable, BlueprintCallable, DisplayName="Set From Name (Server)", Category = "NEXUS|Multiplayer")
+	void Server_SetFromName(const FName& NewValue);
 	
-	UFUNCTION(Server, Reliable, BlueprintCallable, Category = "NEXUS")
-	void SetFromString(const FString& NewValue);
+	UFUNCTION(Server, Reliable, BlueprintCallable, DisplayName="Set From String (Server)", Category = "NEXUS|Multiplayer")
+	void Server_SetFromString(const FString& NewValue);
 
-	UFUNCTION(Server, Reliable, BlueprintCallable, Category = "NEXUS")
-	void SetFromText(const FText& NewValue);
+	UFUNCTION(Server, Reliable, BlueprintCallable, DisplayName="Set From Text (Server)", Category = "NEXUS|Multiplayer")
+	void Server_SetFromText(const FText& NewValue);
+	
+	UFUNCTION(BlueprintCallable, DisplayName="Set From Name (World Authority)", Category = "NEXUS|Multiplayer")
+	void WorldAuthority_SetFromName(const FName& NewValue);
+
+	UFUNCTION(BlueprintCallable, DisplayName="Set From String (World Authority)", Category = "NEXUS|Multiplayer")
+	void WorldAuthority_SetFromString(const FString& NewValue);
+
+	UFUNCTION(BlueprintCallable, DisplayName="Set From Text (World Authority)",  Category = "NEXUS|Multiplayer")
+	void WorldAuthority_SetFromText(const FText& NewValue);
 
 	UPROPERTY(BlueprintAssignable)
 	FOnTextChanged OnTextChanged;